relocated components folder into app
