+++ app/components/HomeHero/HeroCarousel/HeroCarouselHelpers.ts
