+++ app/components/HomeHero/HomeHero.tsx
@@ -1,6 +1,6 @@
 import { client } from "@/sanity/client";
 
-import HeroCarousel from "@/components/HomeHero/HeroCarousel/HeroCarousel";
+import HeroCarousel from "@/app/components/HomeHero/HeroCarousel/HeroCarousel";
 
 export default async function HomeHero() {
   const bands = await client.fetch<Band[]>(`*[_type == "band"]`);
